requirements.txt
flask
pyjwt
PyMySQL
scrypt
flask_sqlalchemy
flask_script
flask_migrate
python-dotenv
pyyaml